requirements.txt
Flask==2.1.0
PyPDF2>=3.0.0
openpyxl==3.1.2
python-docx==0.8.10
beautifulsoup4==4.9.3
requests==2.24.0
python-pptx==0.6.18
youtube_transcript_api==0.6.1
moviepy==1.0.3
pydub==0.25.1
numpy==1.25.2